refactor qc gadgets
