+++ Cargo.toml
@@ -26,9 +26,9 @@ digest = { version = "0.10" }
 displaydoc = { version = "0.2.3", default-features = false }
 ethereum-types = { version = "0.14.1", features = ["impl-serde"] }
 generic-array = "0.14.7"
-jf-primitives = { git = "https://github.com/espressosystems/jellyfish" }
-jf-relation = { git = "https://github.com/espressosystems/jellyfish" }
-jf-utils = { git = "https://github.com/espressosystems/jellyfish" }
+jf-primitives = { git = "https://github.com/espressosystems/jellyfish", branch = "snarked-qc-helper" } # TODO: remove branch after jellyfish update
+jf-relation = { git = "https://github.com/espressosystems/jellyfish", branch = "snarked-qc-helper" } # TODO: remove branch after jellyfish update
+jf-utils = { git = "https://github.com/espressosystems/jellyfish", branch = "snarked-qc-helper" } # TODO: remove branch after jellyfish update
 serde = { version = "1.0", default-features = false, features = ["derive", "rc"] }
 sha3 = "0.10.7"
 tagged-base64 = { git = "https://github.com/espressosystems/tagged-base64", tag = "0.3.0" }
@@ -37,7 +37,7 @@ typenum = { version = "1.16.0" }
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
-jf-primitives = { git = "https://github.com/espressosystems/jellyfish", features = ["test-srs"] }
+jf-primitives = { git = "https://github.com/espressosystems/jellyfish", features = ["test-srs"], branch = "snarked-qc-helper" } # TODO: remove branch after jellyfish update
 sha2 = { version = "0.10" }
 
 [[bench]]
